Connection RNSA data 2 - modification w Lucas
