+++ app/src/main/java/com/example/mint/controller/MapActivity.java
@@ -631,4 +631,4 @@ public class MapActivity<pollen_alert_count, pollen_count, pollen> extends AppCo
             dialog.show();
         }
 
-    }
+}
