+++ app/src/main/java/com/example/mint/model/CustomPopup.java
@@ -22,7 +22,7 @@ public class CustomPopup extends Dialog {
        setContentView(R.layout.popup_pollen);
        this.title = "Alerte Pollen";
        this.subTitle = "Alerte pollen au niveau 4";
-       this.titleView = findViewById(R.id.pollen_alert1);
+       this.titleView = findViewById(R.id.pollen_alert_title);
        //this.titleView = findViewById(R.id.pollen_alert2);
 
    }
