+++ app/src/main/java/com/example/mint/model/MyHttpUtils.java
@@ -12,8 +12,7 @@ public class MyHttpUtils {
     public static String getDataHttpUriConnection(String uri){
         try {
             URL url = new URL("http://51.77.201.227:100/pickdate/noemie/12_25");
-            HttpURLConnection urlConnection = (HttpURLConnection) url
-                    .openConnection();
+            HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
             String donneesPollen = inputStreamToString(urlConnection.getInputStream());
             return donneesPollen;
         } catch (IOException e) {
